Redo distribution and return period interface
